+++ mllc/src/mono.rs
@@ -220,6 +220,7 @@ impl Monomorphizer {
             "bsGetU16LE", "bsGetU32LE", "bsGetI8", "bsGetI16LE", "bsPutI16LE",
             "runST", "newSTArray", "readSTArray", "writeSTArray",
             "modifySTArray", "stArrayLength", "newSTArrayFromList", "stArrayToList",
+            "newIORef", "readIORef", "writeIORef", "modifyIORef", "modifyIORef'",
         ] {
             builtins.insert(name.to_string());
         }